pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-soc"
version = "0.1.0"
description = "Agentic Security Operations Center - AI-powered SOC automation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "ZTE Agentic SOC Team"}
]
keywords = ["security", "soc", "ai", "agents", "azure", "sentinel"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Microsoft Foundry and Agent Framework
    "azure-ai-projects>=1.0.0b2",
    "agent-framework>=0.1.0",
    
    # Azure SDK Core
    "azure-identity>=1.15.0",
    "azure-core>=1.29.0",
    
    # Azure Services
    "azure-cosmos>=4.5.0",
    "azure-storage-blob>=12.19.0",
    "azure-eventhub>=5.11.0",
    "azure-keyvault-secrets>=4.7.0",
    
    # Monitoring and Observability
    "azure-monitor-opentelemetry>=1.2.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "structlog>=23.2.0",
    
    # Data Processing
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pandas>=2.1.0",
    "polars>=0.20.0",
    
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    
    # CLI
    "click>=8.1.7",
    "rich>=13.7.0",
    "typer>=0.9.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = [
    # Linting and Formatting
    "black>=24.1.0",
    "pylint>=3.0.3",
    "mypy>=1.8.0",
    "ruff>=0.1.13",
    
    # Testing
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    
    # Security
    "bandit>=1.7.6",
    "safety>=3.0.0",
    
    # Type Stubs
    "types-requests>=2.31.0",
    "pandas-stubs>=2.1.0",
]

[project.scripts]
asoc = "src.demo.cli:main"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | build
  | dist
)/
'''

[tool.pylint.main]
max-line-length = 100
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "R0903",  # too-few-public-methods
    "R0913",  # too-many-arguments
    "W0212",  # protected-access
]
good-names = ["i", "j", "k", "ex", "id", "df", "_"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "S",   # flake8-bandit
]
ignore = [
    "S101",  # assert-used (we use asserts in tests)
    "S104",  # hardcoded-bind-all-interfaces
    "S106",  # hardcoded-password-func-arg
]

[tool.ruff.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
